refactor(CO-3211): update attachments type imports to use specific paths for better clarity
